move `build_agg_call_from_prost` to dedicated file
