+++ src/stream/src/from_proto/global_simple_agg.rs
@@ -14,6 +14,7 @@
 
 //! Streaming Aggregators
 
+use super::agg_call::build_agg_call_from_prost;
 use super::*;
 use crate::executor::aggregation::{generate_state_tables_from_proto, AggCall};
 use crate::executor::GlobalSimpleAggExecutor;
