+++ src/stream/src/from_proto/local_simple_agg.rs
@@ -12,14 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::convert::TryFrom;
-use std::sync::Arc;
-
-use risingwave_common::types::DataType;
-use risingwave_expr::expr::{build_from_prost, AggKind};
-
+use super::agg_call::build_agg_call_from_prost;
 use super::*;
-use crate::executor::aggregation::{AggArgs, AggCall};
+use crate::executor::aggregation::AggCall;
 use crate::executor::LocalSimpleAggExecutor;
 
 pub struct LocalSimpleAggExecutorBuilder;
@@ -46,34 +41,4 @@ impl ExecutorBuilder for LocalSimpleAggExecutorBuilder {
         )?
         .boxed())
     }
-}
-
-pub fn build_agg_call_from_prost(
-    append_only: bool,
-    agg_call_proto: &risingwave_pb::expr::AggCall,
-) -> Result<AggCall> {
-    let args = match &agg_call_proto.get_args()[..] {
-        [] => AggArgs::None,
-        [arg] => AggArgs::Unary(
-            DataType::from(arg.get_type()?),
-            arg.get_input()?.column_idx as usize,
-        ),
-        _ => {
-            return Err(RwError::from(ErrorCode::NotImplemented(
-                "multiple aggregation args".to_string(),
-                None.into(),
-            )))
-        }
-    };
-    let filter = match agg_call_proto.filter {
-        Some(ref prost_filter) => Some(Arc::from(build_from_prost(prost_filter)?)),
-        None => None,
-    };
-    Ok(AggCall {
-        kind: AggKind::try_from(agg_call_proto.get_type()?)?,
-        args,
-        return_type: DataType::from(agg_call_proto.get_return_type()?),
-        append_only,
-        filter,
-    })
 }
